+++ pg-common/src/main/java/fr/theorozier/procgen/common/world/util/DummyWorld.java
@@ -69,7 +69,7 @@ public class DummyWorld implements WorldAccessor {
 	}
 
 	@Override
-	public int getVerticalChunkCount() {
+	public short getVerticalChunkCount() {
 		return 1;
 	}
